domain_checker/requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
python-dotenv==1.0.0
urllib3==2.1.0
python-telegram-bot==21.0.1

